remove old course specifics from adapt logic (#11)
Remove logic which is specified to a certain course

Not every course has pre-assessment problems, that is why first problem should be chosen
from main scope if pre-assessment problems are absent.

The same decision should be done to attention question, which could be absent in course or
mistake could be made during answer and student won't be able to continue. It looks incorrect.